Define the public exports
* make all internal modules private
* remove visibility from types that should not be pub
